fix enemy and player
-renamed EnemyController to EnemyManager
-moved enemy detection function from EnemyManager to EnemyNode class
-fixed player’s animation speed not being updated in touchesBegan
