+++ HachitaiGame/Obstacle.swift
@@ -19,7 +19,7 @@ class ObstacleSpriteNode: SKSpriteNode {
     }
     
     required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
+        super.init(coder: aDecoder)
     }
 }
