+++ HachitaiGame/Player.swift
@@ -12,7 +12,7 @@ import GameKit
 class PlayerNode: SKSpriteNode {
     required init() {
         super.init(texture: SKTexture(image: #imageLiteral(resourceName: "player2")), color: UIColor.white ,size: CGSize(width: 20, height: 20))
-        self.addChild(Utility.makeArrow())
+
         self.physicsBody = SKPhysicsBody(rectangleOf: self.size)
         self.physicsBody?.affectedByGravity = false
         self.physicsBody?.isDynamic = true
@@ -21,7 +21,7 @@ class PlayerNode: SKSpriteNode {
     }
     
     required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
+        super.init(coder: aDecoder)
     }
